third_party/blink/web_tests/wpt_internal/digital-credentials/digital-credential-user-agent-allows-protocol.https.html

<!DOCTYPE html>
<title>Tests for DigitalCredential's userAgentAllowsProtocol() static method.</title>
<link rel="help" href="https://github.com/w3c-fedid/digital-credentials/pull/221" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body></body>
<script type="module">

// Internal test because webkit doesn't support openid4vp yet.
promise_test(async (t) => {
  const openid4vp_identifiers = ["openid4vp", "openid4vp-v1-signed", "openid4vp-v1-unsigned"];
    for (const openid4vp_identifier of openid4vp_identifiers) {
      const result = DigitalCredential.userAgentAllowsProtocol(openid4vp_identifier);
      assert_equals(result, true, `Incorrect result for ${openid4vp_identifier}.`);
    }
  }, "User agent supports openid4vp protocol.");

  promise_test(async (t) => {
      assert_equals(DigitalCredential.userAgentAllowsProtocol("org-iso-mdoc"), true);
  }, "User agent supports iso protocol.");

</script>
